requirements.txt
ffmpeg-python
imageio
imageio-ffmpeg
matplotlib
numpy
pandas
python-dateutil
pytz
PyYAML
scikit-image
scikit-learn
scipy
torch
torchvision
tqdm
opencv-python
uvicorn
fastapi
websockets
Cython
